flask local state + app follow/unfollow
+ following + refactoring of code --> note: not tested everything yet so some code paths may be broken by refactoring – TODO: test this
