+++ examples/flask-with-lenspy/templates/main.html
@@ -20,7 +20,7 @@
 		<script src="https://cdn.jsdelivr.net/gh/ethereum/web3.js/dist/web3.min.js"></script>
 		<script src="https://cdn.ethers.io/lib/ethers-5.2.umd.min.js" type="application/javascript"></script>
 		<script src='https://cdnjs.cloudflare.com/ajax/libs/jquery/3.6.1/jquery.min.js'></script>
-		<script type="module" src="{{url_for('static',filename='js/script.js') }}"></script>
+		<script type="module" src="{{url_for('static',filename='js/main.js') }}"></script>
 		{% block extrajs %}{% endblock %}
 		<!-- <script src="https://cdn.jsdelivr.net/npm/web3@1.8.2/dist/web3.min.js"></script> -->
 	</head>
