+++ examples/flask-with-lenspy/templates/profile.html
@@ -1,4 +1,4 @@
-{% extends 'base.html' %}
+{% extends 'main.html' %}
 
 {% block content %}
 	<div id="view-profile" class="content-section hide">
@@ -25,7 +25,7 @@
 					<div id="follow-profile" class="button content-button">
 						<p>Follow</p>
 					</div>
-					<div id="unfollow-profile" class="button content-button done">
+					<div id="unfollow-profile" class="button content-button">
 						<p>Unfollow</p>
 					</div>
 				</div>
