+++ examples/flask-with-lenspy/templates/publication.html
@@ -1,4 +1,4 @@
-{% extends 'base.html' %}
+{% extends 'main.html' %}
 
 {% block content %}
 	<div id="view-publication" class="content-section hide">
