+++ examples/flask-with-lenspy/templates/search-profile.html
@@ -1,4 +1,4 @@
-{% extends 'base.html' %}
+{% extends 'main.html' %}
 
 {% block content %}
 	<div id="profile-search" class="content-section hide">
